requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.18
botocore==1.37.18
branca==0.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.3
comtypes==1.4.10
contourpy==1.3.1
croniter==6.0.0
cryptography==45.0.5
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
email-validator==2.3.0
emoji==2.15.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.120.4
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
filelock==3.18.0
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
folium==0.19.5
fonttools==4.55.3
frozenlist==1.8.0
fsspec==2025.7.0
future==1.0.0
gast==0.6.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-genai==1.7.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
h2==4.3.0
h5py==3.15.1
hf-xet==1.2.0
holidays==0.76
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.context==6.0.1
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.4.2
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.12.0
kiwisolver==1.4.8
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.2
langchain-gemini==0.1.1
langchain-google-genai==3.0.0
langchain-openai==1.0.1
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langsmith==0.4.38
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.6.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.2.0
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.4.0
oauth2client==4.1.3
oauthlib==3.2.2
openai==2.6.1
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.3
parso==0.8.4
pdf_processor==0.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.1.0
pipwin==0.5.2
platformdirs==4.3.8
playsound==1.2.2
pocketsphinx==5.0.4
portalocker==3.2.0
prompt_toolkit==3.0.51
propcache==0.4.1
prophet==1.1.7
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyjokes==0.8.3
pyjsparser==2.7.1
pymongo==4.13.2
PyMsgBox==1.0.9
PyMuPDF==1.26.3
pyparsing==3.2.1
pypdf==6.1.3
pypdfium2==4.30.0
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
pywin32==309
PyYAML==6.0.2
pyzmq==27.0.1
qdrant-client==1.15.1
ratelim==0.1.6
redis==7.0.1
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.3
rpds-py==0.30.0
rq==2.6.0
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
sambanova==1.3.0
scikit-learn==1.7.2
scipy==1.15.2
seaborn==0.13.2
selenium==4.30.0
sentence-transformers==5.0.0
sentry-sdk==2.43.0
setuptools==77.0.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.14.2
SQLAlchemy==2.0.39
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
stanio==0.5.1
starlette==0.49.3
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
trio==0.29.0
trio-websocket==0.12.2
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==6.0.1
urlextract==1.9.0
urllib3==2.3.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wolframalpha==5.1.3
wordcloud==1.9.5
wrapt==2.0.1
wsproto==1.2.0
xgboost==3.0.5
xlrd==2.0.2
xmltodict==0.14.2
xyzservices==2025.1.0
yarl==1.22.0
zstandard==0.25.0
